feat: scroll based animation
