+++ src/components/Wallet/VerticalSlide/styles.module.css
@@ -1,3 +1,18 @@
+.sectionContainer {
+  height: 240vh;
+}
+
+.stickyContainer {
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  position: sticky;
+  top: 0;
+  overflow: hidden;
+}
+
 .title {
   text-align: center;
   max-width: 650px;
